+++ nflow-netty/src/test/java/io/nflow/netty/StartNflowTest.java
@@ -1,9 +1,12 @@
 package io.nflow.netty;
 
 import static io.nflow.rest.v1.ResourcePaths.NFLOW_WORKFLOW_DEFINITION_PATH;
+import static io.nflow.rest.v1.ResourcePaths.NFLOW_WORKFLOW_INSTANCE_PATH;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
 
 import java.util.HashMap;
@@ -20,8 +23,12 @@ import com.fasterxml.jackson.databind.JsonNode;
 
 public class StartNflowTest {
 
+  private static final String DEFAULT_LOCALHOST_SERVER_PORT = "7500";
+  private static final String DEFAULT_LOCALHOST_SERVER_ADDRESS = "http://localhost:" + DEFAULT_LOCALHOST_SERVER_PORT;
+
   public class TestApplicationListener implements ApplicationListener<ApplicationContextEvent> {
     public ApplicationContextEvent applicationContextEvent;
+
     @Override
     public void onApplicationEvent(ApplicationContextEvent event) {
       applicationContextEvent = event;
@@ -44,7 +51,7 @@ public class StartNflowTest {
     ApplicationContext ctx = startNflow.startNetty(7500, "local", "", properties);
 
     assertNotNull(testListener.applicationContextEvent);
-    assertEquals("7500", ctx.getEnvironment().getProperty("port"));
+    assertEquals(DEFAULT_LOCALHOST_SERVER_PORT, ctx.getEnvironment().getProperty("port"));
     assertEquals("local", ctx.getEnvironment().getProperty("env"));
     assertEquals("externallyDefinedExecutorGroup", ctx.getEnvironment().getProperty("nflow.executor.group"));
     assertEquals("true", ctx.getEnvironment().getProperty("nflow.db.create_on_startup"));
@@ -52,14 +59,27 @@ public class StartNflowTest {
     assertEquals("true", ctx.getEnvironment().getProperty("nflow.autoinit"));
 
     smokeTestRestApi(restApiPrefix);
+    smokeTestRestApiErrorHandling(restApiPrefix);
   }
 
   private void smokeTestRestApi(String restApiPrefix) {
-    WebClient client = WebClient.builder().baseUrl("http://localhost:7500").build();
-    ClientResponse response = client.get().uri(restApiPrefix + NFLOW_WORKFLOW_DEFINITION_PATH).exchange().block();
+    ClientResponse response = getFromDefaultServer(restApiPrefix + NFLOW_WORKFLOW_DEFINITION_PATH);
     assertEquals(OK, response.statusCode());
     JsonNode responseBody = response.bodyToMono(JsonNode.class).block();
     assertTrue(responseBody.isArray());
   }
 
+  // Smoke test for io.nflow.rest.v1.springweb.SpringWebResource#handleExceptions
+  private void smokeTestRestApiErrorHandling(String restApiPrefix) {
+    ClientResponse response = getFromDefaultServer(restApiPrefix + NFLOW_WORKFLOW_INSTANCE_PATH + "/id/0213132");
+    assertEquals(NOT_FOUND, response.statusCode());
+    JsonNode responseBody = response.bodyToMono(JsonNode.class).block();
+    assertNotNull(responseBody);
+    assertFalse(responseBody.isEmpty());
+  }
+
+  private ClientResponse getFromDefaultServer(String url) {
+    WebClient client = WebClient.builder().baseUrl(DEFAULT_LOCALHOST_SERVER_ADDRESS).build();
+    return client.get().uri(url).exchange().block();
+  }
 }