+++ nflow-tests/src/main/java/io/nflow/tests/demo/workflow/TestCronWorkflow.java
@@ -1,6 +1,8 @@
 package io.nflow.tests.demo.workflow;
 
-import static io.nflow.engine.workflow.definition.NextAction.moveToState;
+import static io.nflow.engine.workflow.definition.NextAction.moveToStateAfter;
+import static io.nflow.tests.demo.workflow.DemoWorkflow.DEMO_WORKFLOW_TYPE;
+import static org.joda.time.DateTime.now;
 import static org.joda.time.Period.hours;
 
 import org.joda.time.DateTime;
@@ -10,6 +12,7 @@ import io.nflow.engine.workflow.curated.CronWorkflow;
 import io.nflow.engine.workflow.definition.NextAction;
 import io.nflow.engine.workflow.definition.StateExecution;
 import io.nflow.engine.workflow.definition.WorkflowSettings.Builder;
+import io.nflow.engine.workflow.instance.WorkflowInstance;
 
 @Component
 public class TestCronWorkflow extends CronWorkflow {
@@ -19,8 +22,10 @@ public class TestCronWorkflow extends CronWorkflow {
     super(TYPE, new Builder().setHistoryDeletableAfter(hours(1)).setDeleteHistoryCondition(() -> true).build());
   }
 
-  public NextAction doWork(@SuppressWarnings("unused") StateExecution execution) {
-    return moveToState(SCHEDULE, "ok");
+  public NextAction doWork(StateExecution execution) {
+    WorkflowInstance childWorkflow = new WorkflowInstance.Builder().setType(DEMO_WORKFLOW_TYPE).build();
+    execution.addChildWorkflows(childWorkflow);
+    return moveToStateAfter(WAIT_FOR_WORK_TO_FINISH, now().plusMinutes(1), "Work delegated to child workflow");
   }
 
   @Override
@@ -31,5 +36,10 @@ public class TestCronWorkflow extends CronWorkflow {
   @Override
   protected boolean handleFailureImpl(StateExecution execution) {
     return super.handleFailureImpl(execution);
+  }
+
+  @Override
+  protected DateTime waitForWorkToFinishImpl(StateExecution execution) {
+    return super.waitForWorkToFinishImpl(execution);
   }
 }