+++ nflow-tests/src/test/java/io/nflow/tests/CronWorkflowTest.java
@@ -3,6 +3,7 @@ package io.nflow.tests;
 import static io.nflow.engine.workflow.curated.CronWorkflow.DO_WORK;
 import static io.nflow.engine.workflow.curated.CronWorkflow.FAILED;
 import static io.nflow.engine.workflow.curated.CronWorkflow.SCHEDULE;
+import static io.nflow.engine.workflow.curated.CronWorkflow.WAIT_FOR_WORK_TO_FINISH;
 import static io.nflow.tests.demo.workflow.TestCronWorkflow.TYPE;
 import static java.util.Collections.singletonMap;
 import static java.util.concurrent.TimeUnit.SECONDS;
@@ -55,8 +56,8 @@ public class CronWorkflowTest extends AbstractNflowTest {
 
   @Test
   @Order(2)
-  public void letItRunFor5Seconds() throws InterruptedException {
-    SECONDS.sleep(5);
+  public void letItRunForTenSeconds() throws InterruptedException {
+    SECONDS.sleep(10);
   }
 
   @Test
@@ -64,8 +65,10 @@ public class CronWorkflowTest extends AbstractNflowTest {
   public void verifyItHasRunPeriodically() {
     List<Action> actions = getWorkflowInstance(resp.id).actions;
     long scheduleActions = actions.stream().filter(a -> SCHEDULE.name().equals(a.state)).count();
+    long waitActions = actions.stream().filter(a -> WAIT_FOR_WORK_TO_FINISH.name().equals(a.state)).count();
     long doWorkActions = actions.stream().filter(a -> DO_WORK.name().equals(a.state)).count();
     assertThat(scheduleActions, is(greaterThanOrEqualTo(1L)));
+    assertThat(waitActions, is(greaterThanOrEqualTo(1L)));
     assertThat(doWorkActions, is(greaterThanOrEqualTo(1L)));
   }